Move D&D logic to handlers on elements
